Struct aws_sdk_iotsitewise::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSIoTSiteWise.

This client allows ergonomic access to a AWSIoTSiteWise-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_assets(&self) -> AssociateAssets<C, M, R>

pub fn batch_associate_project_assets(
&self
) -> BatchAssociateProjectAssets<C, M, R>

pub fn batch_disassociate_project_assets(
&self
) -> BatchDisassociateProjectAssets<C, M, R>

pub fn batch_put_asset_property_value(
&self
) -> BatchPutAssetPropertyValue<C, M, R>

pub fn create_access_policy(&self) -> CreateAccessPolicy<C, M, R>

pub fn create_asset(&self) -> CreateAsset<C, M, R>

pub fn create_asset_model(&self) -> CreateAssetModel<C, M, R>

pub fn create_dashboard(&self) -> CreateDashboard<C, M, R>

pub fn create_gateway(&self) -> CreateGateway<C, M, R>

pub fn create_portal(&self) -> CreatePortal<C, M, R>

pub fn create_project(&self) -> CreateProject<C, M, R>

pub fn delete_access_policy(&self) -> DeleteAccessPolicy<C, M, R>

pub fn delete_asset(&self) -> DeleteAsset<C, M, R>

pub fn delete_asset_model(&self) -> DeleteAssetModel<C, M, R>

pub fn delete_dashboard(&self) -> DeleteDashboard<C, M, R>

pub fn delete_gateway(&self) -> DeleteGateway<C, M, R>

pub fn delete_portal(&self) -> DeletePortal<C, M, R>

pub fn delete_project(&self) -> DeleteProject<C, M, R>

pub fn describe_access_policy(&self) -> DescribeAccessPolicy<C, M, R>

pub fn describe_asset(&self) -> DescribeAsset<C, M, R>

pub fn describe_asset_model(&self) -> DescribeAssetModel<C, M, R>

pub fn describe_asset_property(&self) -> DescribeAssetProperty<C, M, R>

pub fn describe_dashboard(&self) -> DescribeDashboard<C, M, R>

pub fn describe_default_encryption_configuration(
&self
) -> DescribeDefaultEncryptionConfiguration<C, M, R>

pub fn describe_gateway(&self) -> DescribeGateway<C, M, R>

pub fn describe_gateway_capability_configuration(
&self
) -> DescribeGatewayCapabilityConfiguration<C, M, R>

pub fn describe_logging_options(&self) -> DescribeLoggingOptions<C, M, R>

pub fn describe_portal(&self) -> DescribePortal<C, M, R>

pub fn describe_project(&self) -> DescribeProject<C, M, R>

pub fn describe_storage_configuration(
&self
) -> DescribeStorageConfiguration<C, M, R>

pub fn disassociate_assets(&self) -> DisassociateAssets<C, M, R>

pub fn get_asset_property_aggregates(
&self
) -> GetAssetPropertyAggregates<C, M, R>

pub fn get_asset_property_value(&self) -> GetAssetPropertyValue<C, M, R>

pub fn get_asset_property_value_history(
&self
) -> GetAssetPropertyValueHistory<C, M, R>

pub fn get_interpolated_asset_property_values(
&self
) -> GetInterpolatedAssetPropertyValues<C, M, R>

pub fn list_access_policies(&self) -> ListAccessPolicies<C, M, R>

pub fn list_asset_models(&self) -> ListAssetModels<C, M, R>

pub fn list_asset_relationships(&self) -> ListAssetRelationships<C, M, R>

pub fn list_assets(&self) -> ListAssets<C, M, R>

pub fn list_associated_assets(&self) -> ListAssociatedAssets<C, M, R>

pub fn list_dashboards(&self) -> ListDashboards<C, M, R>

pub fn list_gateways(&self) -> ListGateways<C, M, R>

pub fn list_portals(&self) -> ListPortals<C, M, R>

pub fn list_project_assets(&self) -> ListProjectAssets<C, M, R>

pub fn list_projects(&self) -> ListProjects<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn put_default_encryption_configuration(
&self
) -> PutDefaultEncryptionConfiguration<C, M, R>

pub fn put_logging_options(&self) -> PutLoggingOptions<C, M, R>

pub fn put_storage_configuration(&self) -> PutStorageConfiguration<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_access_policy(&self) -> UpdateAccessPolicy<C, M, R>

pub fn update_asset(&self) -> UpdateAsset<C, M, R>

pub fn update_asset_model(&self) -> UpdateAssetModel<C, M, R>

pub fn update_asset_property(&self) -> UpdateAssetProperty<C, M, R>

pub fn update_dashboard(&self) -> UpdateDashboard<C, M, R>

pub fn update_gateway(&self) -> UpdateGateway<C, M, R>

pub fn update_gateway_capability_configuration(
&self
) -> UpdateGatewayCapabilityConfiguration<C, M, R>

pub fn update_portal(&self) -> UpdatePortal<C, M, R>

pub fn update_project(&self) -> UpdateProject<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self