Module aws_sdk_iotsitewise::error

Modules

conflicting_operation_exception

See ConflictingOperationException

internal_failure_exception

See InternalFailureException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

too_many_tags_exception

See TooManyTagsException

unauthorized_exception

See UnauthorizedException

Structs

AssociateAssetsError

BatchAssociateProjectAssetsError

BatchDisassociateProjectAssetsError

BatchPutAssetPropertyValueError

ConflictingOperationException

Your request has conflicting operations. This can occur if you're trying to perform more

CreateAccessPolicyError

CreateAssetError

CreateAssetModelError

CreateDashboardError

CreateGatewayError

CreatePortalError

CreateProjectError

DeleteAccessPolicyError

DeleteAssetError

DeleteAssetModelError

DeleteDashboardError

DeleteGatewayError

DeletePortalError

DeleteProjectError

DescribeAccessPolicyError

DescribeAssetError

DescribeAssetModelError

DescribeAssetPropertyError

DescribeDashboardError

DescribeDefaultEncryptionConfigurationError

DescribeGatewayCapabilityConfigurationError

DescribeGatewayError

DescribeLoggingOptionsError

DescribePortalError

DescribeProjectError

DescribeStorageConfigurationError

DisassociateAssetsError

GetAssetPropertyAggregatesError

GetAssetPropertyValueError

GetAssetPropertyValueHistoryError

GetInterpolatedAssetPropertyValuesError

InternalFailureException

IoT SiteWise can't process your request right now. Try again later.

InvalidRequestException

The request isn't valid. This can occur if your request contains malformed JSON or

LimitExceededException

You've reached the limit for a resource. For example, this can occur if you're trying to

ListAccessPoliciesError

ListAssetModelsError

ListAssetRelationshipsError

ListAssetsError

ListAssociatedAssetsError

ListDashboardsError

ListGatewaysError

ListPortalsError

ListProjectAssetsError

ListProjectsError

ListTagsForResourceError

PutDefaultEncryptionConfigurationError

PutLoggingOptionsError

PutStorageConfigurationError

ResourceAlreadyExistsException

The resource already exists.

ResourceNotFoundException

The requested resource can't be found.

ServiceUnavailableException

The requested service is unavailable.

TagResourceError

ThrottlingException

Your request exceeded a rate limit. For example, you might have exceeded the number of

TooManyTagsException

You've reached the limit for the number of tags allowed for a resource. For more

UnauthorizedException

You are not authorized.

UntagResourceError

UpdateAccessPolicyError

UpdateAssetError

UpdateAssetModelError

UpdateAssetPropertyError

UpdateDashboardError

UpdateGatewayCapabilityConfigurationError

UpdateGatewayError

UpdatePortalError

UpdateProjectError

Enums

AssociateAssetsErrorKind

BatchAssociateProjectAssetsErrorKind

BatchDisassociateProjectAssetsErrorKind

BatchPutAssetPropertyValueErrorKind

CreateAccessPolicyErrorKind

CreateAssetErrorKind

CreateAssetModelErrorKind

CreateDashboardErrorKind

CreateGatewayErrorKind

CreatePortalErrorKind

CreateProjectErrorKind

DeleteAccessPolicyErrorKind

DeleteAssetErrorKind

DeleteAssetModelErrorKind

DeleteDashboardErrorKind

DeleteGatewayErrorKind

DeletePortalErrorKind

DeleteProjectErrorKind

DescribeAccessPolicyErrorKind

DescribeAssetErrorKind

DescribeAssetModelErrorKind

DescribeAssetPropertyErrorKind

DescribeDashboardErrorKind

DescribeDefaultEncryptionConfigurationErrorKind

DescribeGatewayCapabilityConfigurationErrorKind

DescribeGatewayErrorKind

DescribeLoggingOptionsErrorKind

DescribePortalErrorKind

DescribeProjectErrorKind

DescribeStorageConfigurationErrorKind

DisassociateAssetsErrorKind

GetAssetPropertyAggregatesErrorKind

GetAssetPropertyValueErrorKind

GetAssetPropertyValueHistoryErrorKind

GetInterpolatedAssetPropertyValuesErrorKind

ListAccessPoliciesErrorKind

ListAssetModelsErrorKind

ListAssetRelationshipsErrorKind

ListAssetsErrorKind

ListAssociatedAssetsErrorKind

ListDashboardsErrorKind

ListGatewaysErrorKind

ListPortalsErrorKind

ListProjectAssetsErrorKind

ListProjectsErrorKind

ListTagsForResourceErrorKind

PutDefaultEncryptionConfigurationErrorKind

PutLoggingOptionsErrorKind

PutStorageConfigurationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAccessPolicyErrorKind

UpdateAssetErrorKind

UpdateAssetModelErrorKind

UpdateAssetPropertyErrorKind

UpdateDashboardErrorKind

UpdateGatewayCapabilityConfigurationErrorKind

UpdateGatewayErrorKind

UpdatePortalErrorKind

UpdateProjectErrorKind