Module aws_sdk_iotsitewise::model

Modules

access_policy_summary

See AccessPolicySummary

aggregated_value

See AggregatedValue

aggregates

See Aggregates

alarms

See Alarms

asset_composite_model

See AssetCompositeModel

asset_error_details

See AssetErrorDetails

asset_hierarchy

See AssetHierarchy

asset_hierarchy_info

See AssetHierarchyInfo

asset_model_composite_model

See AssetModelCompositeModel

asset_model_composite_model_definition

See AssetModelCompositeModelDefinition

asset_model_hierarchy

See AssetModelHierarchy

asset_model_hierarchy_definition

See AssetModelHierarchyDefinition

asset_model_property

See AssetModelProperty

asset_model_property_definition

See AssetModelPropertyDefinition

asset_model_status

See AssetModelStatus

asset_model_summary

See AssetModelSummary

asset_property

See AssetProperty

asset_property_value

See AssetPropertyValue

asset_relationship_summary

See AssetRelationshipSummary

asset_status

See AssetStatus

asset_summary

See AssetSummary

associated_assets_summary

See AssociatedAssetsSummary

attribute

See Attribute

batch_put_asset_property_error

See BatchPutAssetPropertyError

batch_put_asset_property_error_entry

See BatchPutAssetPropertyErrorEntry

composite_model_property

See CompositeModelProperty

configuration_error_details

See ConfigurationErrorDetails

configuration_status

See ConfigurationStatus

customer_managed_s3_storage

See CustomerManagedS3Storage

dashboard_summary

See DashboardSummary

detailed_error

See DetailedError

error_details

See ErrorDetails

expression_variable

See ExpressionVariable

forwarding_config

See ForwardingConfig

gateway_capability_summary

See GatewayCapabilitySummary

gateway_platform

See GatewayPlatform

gateway_summary

See GatewaySummary

greengrass

See Greengrass

greengrass_v2

See GreengrassV2

group_identity

See GroupIdentity

iam_role_identity

See IamRoleIdentity

iam_user_identity

See IamUserIdentity

identity

See Identity

image

See Image

image_file

See ImageFile

image_location

See ImageLocation

interpolated_asset_property_value

See InterpolatedAssetPropertyValue

logging_options

See LoggingOptions

measurement

See Measurement

measurement_processing_config

See MeasurementProcessingConfig

metric

See Metric

metric_processing_config

See MetricProcessingConfig

metric_window

See MetricWindow

monitor_error_details

See MonitorErrorDetails

multi_layer_storage

See MultiLayerStorage

portal_resource

See PortalResource

portal_status

See PortalStatus

portal_summary

See PortalSummary

project_resource

See ProjectResource

project_summary

See ProjectSummary

property

See Property

property_notification

See PropertyNotification

property_type

See PropertyType

put_asset_property_value_entry

See PutAssetPropertyValueEntry

resource

See Resource

time_in_nanos

See TimeInNanos

transform

See Transform

transform_processing_config

See TransformProcessingConfig

tumbling_window

See TumblingWindow

user_identity

See UserIdentity

variable_value

See VariableValue

variant

See Variant

Structs

AccessPolicySummary

Contains an access policy that defines an identity's access to an IoT SiteWise Monitor

AggregatedValue

Contains aggregated asset property values (for example, average, minimum, and

Aggregates

Contains the (pre-calculated) aggregate values for an asset property.

Alarms

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal.

AssetCompositeModel

Contains information about a composite model in an asset. This object contains the asset's

AssetErrorDetails

Contains error details for the requested associate project asset action.

AssetHierarchy

Describes an asset hierarchy that contains a hierarchy's name and ID.

AssetHierarchyInfo

Contains information about a parent asset and a child asset that are related through an

AssetModelCompositeModel

Contains information about a composite model in an asset model. This object contains the

AssetModelCompositeModelDefinition

Contains a composite model definition in an asset model. This composite model definition

AssetModelHierarchy

Describes an asset hierarchy that contains a hierarchy's name, ID, and child asset model

AssetModelHierarchyDefinition

Contains an asset model hierarchy used in asset model creation. An asset model hierarchy

AssetModelProperty

Contains information about an asset model property.

AssetModelPropertyDefinition

Contains an asset model property definition. This property definition is applied to all

AssetModelStatus

Contains current status information for an asset model. For more information, see Asset and model

AssetModelSummary

Contains a summary of an asset model.

AssetProperty

Contains asset property information.

AssetPropertyValue

Contains asset property value information.

AssetRelationshipSummary

Contains information about assets that are related to one another.

AssetStatus

Contains information about the current status of an asset. For more information, see

AssetSummary

Contains a summary of an asset.

AssociatedAssetsSummary

Contains a summary of an associated asset.

Attribute

Contains an asset attribute property. For more information, see

BatchPutAssetPropertyError

Contains error information from updating a batch of asset property values.

BatchPutAssetPropertyErrorEntry

Contains error information for asset property value entries that are associated with the

CompositeModelProperty

Contains information about a composite model property on an asset.

ConfigurationErrorDetails

Contains the details of an IoT SiteWise configuration error.

ConfigurationStatus

Contains current status information for the configuration.

CustomerManagedS3Storage

Contains information about a customer managed Amazon S3 bucket.

DashboardSummary

Contains a dashboard summary.

DetailedError

Contains detailed error information.

ErrorDetails

Contains the details of an IoT SiteWise error.

ExpressionVariable

Contains expression variable information.

ForwardingConfig

The forwarding configuration for a given property.

GatewayCapabilitySummary

Contains a summary of a gateway capability configuration.

GatewayPlatform

Contains a gateway's platform information.

GatewaySummary

Contains a summary of a gateway.

Greengrass

Contains details for a gateway that runs on IoT Greengrass. To create a gateway that runs on IoT Greengrass,

GreengrassV2

Contains details for a gateway that runs on IoT Greengrass V2. To create a gateway that runs on IoT Greengrass V2,

GroupIdentity

Contains information for a group identity in an access policy.

IamRoleIdentity

Contains information about an Identity and Access Management role. For more information, see IAM roles in the

IamUserIdentity

Contains information about an Identity and Access Management user.

Identity

Contains an identity that can access an IoT SiteWise Monitor resource.

Image

Contains an image that is one of the following:

ImageFile

Contains an image file.

ImageLocation

Contains an image that is uploaded to IoT SiteWise and available at a URL.

InterpolatedAssetPropertyValue

Contains information about an interpolated asset property value.

LoggingOptions

Contains logging options.

Measurement

Contains an asset measurement property. For more information, see

MeasurementProcessingConfig

The processing configuration for the given measurement property.

Metric

Contains an asset metric property. With metrics, you can calculate aggregate functions,

MetricProcessingConfig

The processing configuration for the given metric property.

MetricWindow

Contains a time interval window used for data aggregate computations (for example,

MonitorErrorDetails

Contains IoT SiteWise Monitor error details.

MultiLayerStorage

Contains information about the storage destination.

PortalResource

Identifies an IoT SiteWise Monitor portal.

PortalStatus

Contains information about the current status of a portal.

PortalSummary

Contains a portal summary.

ProjectResource

Identifies a specific IoT SiteWise Monitor project.

ProjectSummary

Contains project summary information.

Property

Contains asset property information.

PropertyNotification

Contains asset property value notification information. When the notification state is enabled, IoT SiteWise publishes property value

PropertyType

Contains a property type, which can be one of attribute,

PutAssetPropertyValueEntry

Contains a list of value updates for an asset property in the list of asset entries

Resource

Contains an IoT SiteWise Monitor resource ID for a portal or project.

TimeInNanos

Contains a timestamp with optional nanosecond granularity.

Transform

Contains an asset transform property. A transform is a one-to-one mapping of a property's

TransformProcessingConfig

The processing configuration for the given transform property.

TumblingWindow

Contains a tumbling window, which is a repeating fixed-sized, non-overlapping, and

UserIdentity

Contains information for a user identity in an access policy.

VariableValue

Identifies a property value used in an expression.

Variant

Contains an asset property value (of a single type only).

Enums

AggregateType

AssetErrorCode

AssetModelState

AssetRelationshipType

AssetState

AuthMode

BatchPutAssetPropertyValueErrorCode

CapabilitySyncStatus

NOTE: CapabilitySyncStatus::Unknown has been renamed to ::UnknownValue.

ComputeLocation

ConfigurationState

DetailedErrorCode

EncryptionType

ErrorCode

ForwardingConfigState

IdentityType

ImageFileType

ListAssetsFilter

LoggingLevel

MonitorErrorCode

Permission

PortalState

PropertyDataType

PropertyNotificationState

Quality

ResourceType

StorageType

TimeOrdering

TraversalDirection

TraversalType