Module aws_sdk_iotsitewise::operation

Structs

AssociateAssets

Associates a child asset with the given parent asset through a hierarchy defined in the

BatchAssociateProjectAssets

Associates a group (batch) of assets with an IoT SiteWise Monitor project.

BatchDisassociateProjectAssets

Disassociates a group (batch) of assets from an IoT SiteWise Monitor project.

BatchPutAssetPropertyValue

Sends a list of asset property values to IoT SiteWise. Each value is a timestamp-quality-value

CreateAccessPolicy

Creates an access policy that grants the specified identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or

CreateAsset

Creates an asset from an existing asset model. For more information, see Creating assets in the

CreateAssetModel

Creates an asset model from specified property and hierarchy definitions. You create

CreateDashboard

Creates a dashboard in an IoT SiteWise Monitor project.

CreateGateway

Creates a gateway, which is a virtual or edge device that delivers industrial data streams

CreatePortal

Creates a portal, which can contain projects and dashboards. IoT SiteWise Monitor uses Amazon Web Services SSO or IAM

CreateProject

Creates a project in the specified portal.

DeleteAccessPolicy

Deletes an access policy that grants the specified identity access to the specified

DeleteAsset

Deletes an asset. This action can't be undone. For more information, see Deleting assets and

DeleteAssetModel

Deletes an asset model. This action can't be undone. You must delete all assets created

DeleteDashboard

Deletes a dashboard from IoT SiteWise Monitor.

DeleteGateway

Deletes a gateway from IoT SiteWise. When you delete a gateway, some of the gateway's files remain

DeletePortal

Deletes a portal from IoT SiteWise Monitor.

DeleteProject

Deletes a project from IoT SiteWise Monitor.

DescribeAccessPolicy

Describes an access policy, which specifies an identity's access to an IoT SiteWise Monitor portal or

DescribeAsset

Retrieves information about an asset.

DescribeAssetModel

Retrieves information about an asset model.

DescribeAssetProperty

Retrieves information about an asset property.

DescribeDashboard

Retrieves information about a dashboard.

DescribeDefaultEncryptionConfiguration

Retrieves information about the default encryption configuration for the Amazon Web Services account in

DescribeGateway

Retrieves information about a gateway.

DescribeGatewayCapabilityConfiguration

Retrieves information about a gateway capability configuration.

DescribeLoggingOptions

Retrieves the current IoT SiteWise logging options.

DescribePortal

Retrieves information about a portal.

DescribeProject

Retrieves information about a project.

DescribeStorageConfiguration

Retrieves information about the storage configuration for IoT SiteWise.

DisassociateAssets

Disassociates a child asset from the given parent asset through a hierarchy defined in the

GetAssetPropertyAggregates

Gets aggregated values for an asset property. For more information, see Querying

GetAssetPropertyValue

Gets an asset property's current value. For more information, see Querying

GetAssetPropertyValueHistory

Gets the history of an asset property's values. For more information, see Querying

GetInterpolatedAssetPropertyValues

Get interpolated values for an asset property for a specified time interval, during a

ListAccessPolicies

Retrieves a paginated list of access policies for an identity (an Amazon Web Services SSO user, an Amazon Web Services SSO

ListAssetModels

Retrieves a paginated list of summaries of all asset models.

ListAssetRelationships

Retrieves a paginated list of asset relationships for an asset. You can use this operation

ListAssets

Retrieves a paginated list of asset summaries.

ListAssociatedAssets

Retrieves a paginated list of associated assets.

ListDashboards

Retrieves a paginated list of dashboards for an IoT SiteWise Monitor project.

ListGateways

Retrieves a paginated list of gateways.

ListPortals

Retrieves a paginated list of IoT SiteWise Monitor portals.

ListProjectAssets

Retrieves a paginated list of assets associated with an IoT SiteWise Monitor project.

ListProjects

Retrieves a paginated list of projects for an IoT SiteWise Monitor portal.

ListTagsForResource

Retrieves the list of tags for an IoT SiteWise resource.

PutDefaultEncryptionConfiguration

Sets the default encryption configuration for the Amazon Web Services account. For more information, see

PutLoggingOptions

Sets logging options for IoT SiteWise.

PutStorageConfiguration

Configures storage settings for IoT SiteWise.

TagResource

Adds tags to an IoT SiteWise resource. If a tag already exists for the resource, this operation

UntagResource

Removes a tag from an IoT SiteWise resource.

UpdateAccessPolicy

Updates an existing access policy that specifies an identity's access to an IoT SiteWise Monitor

UpdateAsset

Updates an asset's name. For more information, see Updating assets and models in the

UpdateAssetModel

Updates an asset model and all of the assets that were created from the model. Each asset

UpdateAssetProperty

Updates an asset property's alias and notification state.

UpdateDashboard

Updates an IoT SiteWise Monitor dashboard.

UpdateGateway

Updates a gateway's name.

UpdateGatewayCapabilityConfiguration

Updates a gateway capability configuration or defines a new capability configuration.

UpdatePortal

Updates an IoT SiteWise Monitor portal.

UpdateProject

Updates an IoT SiteWise Monitor project.